+++ CHANGELOG.rst
@@ -117,7 +117,8 @@
   Example '-o xfail_strict=True'. A complete ini-options can be viewed
   by py.test --help. Thanks `@blueyed`_ and `@fengxx`_ for the PR
 
-*
+* Allow passing a custom debugger class (e.g. ``IPython.core.debugger:Pdb``
+  via ``--pdbcls``).  Thanks to `@anntzer`_ for the PR.
 
 *
 
@@ -270,6 +271,7 @@
 .. _@DRMacIver: https://github.com/DRMacIver
 .. _@RedBeardCode: https://github.com/RedBeardCode
 .. _@Vogtinator: https://github.com/Vogtinator
+.. _@anntzer: https://github.com/anntzer
 .. _@bagerard: https://github.com/bagerard
 .. _@blueyed: https://github.com/blueyed
 .. _@ceridwen: https://github.com/ceridwen